Reshape inverted index to be keyed by name, field
Previously, inverted index was keyed by value, with references to the
matching source data sets and identifiers stored as values:

{
  john: [{id: "1", name: "users"}]
}

This changes it to be keyed by name, field, value. The value changes to
ids for the matching source items

{
  users: {
    name: {
      "john": ["1"]
    }
  }
}

This makes it simpler to search keywords in specific fields of specific
data sets, with the tradeoff of being more difficult to search accross
all fields in all datasets at once.